20230323_0009
- DropEnergy를 관리해줄 오브젝트풀링 클래스 생성 후 구현.
- 오브젝트 풀링 갯수만큼 드랍된 에너지가 많으면 많은 만큼 풀링에 추가.

+++ Source/Lakaya/Public/DropEnergyPool.h
@@ -20,7 +20,7 @@ public:
 	virtual void Tick(float DeltaTime) override;
 
 public:
-	void Initialize(uint8 PoolSize);
+	void Initialize(int32 PoolSize);
 	AIndividualDropEnergy* GetDropEnergy();
 	void ReturnDropEnergy(AIndividualDropEnergy* DropEnergy);
